move mock components to separate file
